+++ realm/realm-library/src/main/java/io/realm/internal/ColumnInfo.java
@@ -16,12 +16,11 @@
 
 package io.realm.internal;
 
-import java.util.Collections;
 import java.util.Map;
 
 import io.realm.exceptions.RealmMigrationNeededException;
 
-public class ColumnInfo {
+public abstract class ColumnInfo implements Cloneable {
     private Map<String, Long> indicesMap;
 
     protected final long getValidColumnIndex(String realmPath, Table table,
@@ -34,11 +33,40 @@ public class ColumnInfo {
         return columnIndex;
     }
 
-    protected final void setIndicesMap(Map<String, Long> indicesMap) {
-        this.indicesMap = Collections.unmodifiableMap(indicesMap);
-    }
-
+    /**
+     * Returns a map from column name to column index.
+     *
+     * @return a map from column name to column index. Do not modify returned map because it may be
+     * shared among other {@link ColumnInfo} instances.
+     */
     public Map<String, Long> getIndicesMap() {
         return indicesMap;
     }
+
+    protected final void setIndicesMap(Map<String, Long> indicesMap) {
+        this.indicesMap = indicesMap;
+    }
+
+    /**
+     * Copies the column index value from other {@link ColumnInfo} object.
+     *
+     * @param other The class of {@code other} must be exactly the same as this instance.
+     *              It must not be {@code null}.
+     * @throws IllegalArgumentException if {@code other} has different class than this.
+     */
+    public abstract void copyColumnInfoFrom(ColumnInfo other);
+
+    /**
+     * Returns a shallow copy of this instance.
+     *
+     * @return shallow copy.
+     */
+    @Override
+    public ColumnInfo clone() {
+        try {
+            return (ColumnInfo) super.clone();
+        } catch (CloneNotSupportedException e) {
+            throw new RuntimeException(e);
+        }
+    };
 }